416673: IParserClass interface registered
