+++ src/EPR.SubsidiaryBulkUpload.Application/Services/IParserClass.cs
@@ -1,15 +1,10 @@
-﻿using EPR.SubsidiaryBulkUpload.Application.DTOs;
+﻿using CsvHelper.Configuration;
+using EPR.SubsidiaryBulkUpload.Application.DTOs;
 
 namespace EPR.SubsidiaryBulkUpload.Application.Services
 {
-    internal interface IParserClass
+    public interface IParserClass
     {
-#pragma warning disable SA1414 // Tuple types in signatures should have element names
-        public (ResponseClass, List<CompaniesHouseCompany>) ParseWithHelper(string filePath);
-#pragma warning restore SA1414 // Tuple types in signatures should have element names
-
-        public List<CompaniesHouseCompany> ParseTests(string filePath);
-
-        // public (ResponseClass, List<CompaniesHouseCompany>) ParseErrors(string filePath);
+        public (ResponseClass ResponseClass, List<CompaniesHouseCompany> CompaniesHouseCompany) ParseWithHelper(Stream stream, CsvConfiguration configuration);
     }
 }
