+++ src/EPR.SubsidiaryBulkUpload.Application/Services/ResponseClass.cs
@@ -1,6 +1,6 @@
 ﻿namespace EPR.SubsidiaryBulkUpload.Application.Services
 {
-    internal class ResponseClass
+    public class ResponseClass
     {
         public bool isDone { get; set; }
