+++ src/EPR.SubsidiaryBulkUpload.Function/Extensions/ConfigurationExtensions.cs
@@ -134,6 +134,7 @@ public static class ConfigurationExtensions
         services.AddTransient<ICompaniesHouseDataProvider, CompaniesHouseDataProvider>();
         services.AddTransient<IRecordExtraction, RecordExtraction>();
         services.AddTransient<ICsvProcessor, CsvProcessor>();
+        services.AddTransient<IParserClass, ParserClass>();
         services.AddTransient<ICompaniesHouseCsvProcessor, CompaniesHouseCsvProcessor>();
         services.AddTransient<ITableStorageProcessor, TableStorageProcessor>();
         services.AddTransient<IAzureStorageTableService, AzureStorageTableService>();
